+++ noodles-sam/examples/sam_reheader_async.rs
@@ -28,7 +28,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut writer = sam::AsyncWriter::new(io::stdout());
     writer.write_header(&header).await?;
 
-    let mut records = reader.records();
+    let mut records = reader.records(&header);
 
     while let Some(record) = records.try_next().await? {
         writer.write_record(&header, &record).await?;