+++ noodles-sam/examples/sam_view.rs
@@ -20,7 +20,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let handle = stdout.lock();
     let mut writer = sam::Writer::new(BufWriter::new(handle));
 
-    for result in reader.records() {
+    for result in reader.records(&header) {
         let record = result?;
         writer.write_record(&header, &record)?;
     }